Mark android.frameworks.vibrator as optional
Not all Android devices have a vibrator. The VibratorControlService
should be an optional service for Android devices. We create a new
manifest without the VibratorControlService to avoid running VTS of
VibratorControlService for those devices. Devices only need to set the
soong config "without_vibrator" to exclude from the VTS test.

Flag: EXEMPT bugfix
Bug: 351256023
Test: Test on a TV device and the VTS test is ignored.
Change-Id: I29c17ebeea5fe190221a490ed4388a33ca7bd950